Add instances for Data.Tree.Tree
Rework of https://github.com/ekmett/recursion-schemes/pull/93
