+++ recursion-schemes.cabal
@@ -46,6 +46,7 @@ library
 
   build-depends:
     base                 >= 4.5     && < 5,
+    containers           >= 0.4.2.1 && < 0.7,
     comonad              >= 4       && < 6,
     free                 >= 4       && < 6,
     transformers         >= 0.3.0.0 && < 1
@@ -73,9 +74,10 @@ library
       nats,
       transformers-compat  >= 0.3     && < 1
 
+  -- Foldable module is first, so cabal repl loads it!
   exposed-modules:
-    Data.Functor.Base
     Data.Functor.Foldable
+    Data.Functor.Base
 
   if flag(template-haskell)
     build-depends:
